Synthesizer: Resample operator first version

+++ Data/JJ.Data.Synthesizer/OperatorType.cs
@@ -10,6 +10,6 @@ namespace JJ.Data.Synthesizer
     {
         public virtual int ID { get; set; }
         public virtual string Name { get; set; }
-        public virtual int SortOrder { get; set; }
+        public virtual float SortOrder { get; set; }
     }
 }
